requirements.txt
transformers
nltk
google-generativeai
python-dotenv
flask
